feature: deprecate pmmv4
