+++ src/signer/pmmv5.ts
@@ -3,10 +3,12 @@ import {
   orderHashUtils,
   SignatureType,
   signatureUtils,
+  SignerType,
 } from '0x-v2-order-utils'
+import * as ethUtils from 'ethereumjs-util'
 import { utils, Wallet } from 'ethers'
 import { BigNumber, orderBNToString } from '../utils'
-import { signWithUserAndFee } from './pmmv4'
+import { ecSignOrderHash } from './ecsign'
 
 // changes of PMMV5
 // - taker address point to PMM contract
@@ -19,6 +21,45 @@ export const generateSaltWithFeeFactor = (feeFactor: number) => {
   return new BigNumber(generatePseudoRandomSalt().toString(16).slice(0, -4) + feeHex.slice(2), 16)
 }
 
+// Signature:
+// +------|---------|---------|---------|---------+
+// |  V   |    R    |    S    |userAddr |feeFactor|
+// +------|---------|---------|---------|---------+
+export function signWithUserAndFee(
+  signer: Wallet,
+  orderHash: string,
+  userAddr: string,
+  feeFactor: number
+) {
+  const hash = ethUtils.bufferToHex(
+    Buffer.concat([
+      ethUtils.toBuffer(orderHash),
+      ethUtils.toBuffer(userAddr.toLowerCase()),
+      ethUtils.toBuffer(feeFactor > 255 ? feeFactor : [0, feeFactor]),
+    ])
+  )
+
+  const signature = ecSignOrderHash(
+    signer.privateKey.slice(2),
+    hash,
+    signer.address,
+    SignerType.Default
+  )
+
+  const walletSign = ethUtils.bufferToHex(
+    Buffer.concat([
+      ethUtils.toBuffer(signature).slice(0, 65),
+      ethUtils.toBuffer(userAddr.toLowerCase()),
+      ethUtils.toBuffer(feeFactor > 255 ? feeFactor : [0, feeFactor]),
+    ])
+  )
+  return walletSign
+}
+
+// Signature:
+// +------|---------|---------|---------+
+// |  v   |    R    |    S    | type(3) |
+// +------|---------|---------|---------+
 async function signByEOA(orderHash: string, wallet: Wallet): Promise<string> {
   const hashArray = utils.arrayify(orderHash)
   let signature = await wallet.signMessage(hashArray)
@@ -29,6 +70,11 @@ async function signByEOA(orderHash: string, wallet: Wallet): Promise<string> {
   return signatureUtils.convertToSignatureWithType(signature, SignatureType.EthSign)
 }
 
+// For V4 Maket Maker Proxy (MMP)
+// Signature:
+// +------|---------|---------|---------|---------|---------+
+// |  V   |    R    |    S    |userAddr |feeFactor| type(4) |
+// +------|---------|---------|---------|---------|---------+
 function signByMMPSigner(
   orderHash: string,
   userAddr: string,
